Update Chockalet Problem .cpp

+++ Chockalet Problem .cpp
@@ -3,7 +3,7 @@ public:
     long long findMinDiff(vector<long long> a, long long n, long long m){
     //code
            sort(a.begin(),a.end());
-           long long mi-diff=INT_MAX;
+           long long mi-diff=LLONG_MAX;
            long long diff;
            for(long long  i=0;(m+i-1)<n;i++){
                diff=a[m+i-1]-a[i];
